Update app.component.css
Hey, the reason the image doesn't cover the entire page is because the content makes the page larger so the body scrolls.  This is fixed by adding overflow auto to the content div.  After this we need to add position fixed and set the top to the height of the navbar which is 56px including padding, then bottom 0 and set the overflow to auto.